added sizing support

+++ Editor/MonoBehaviour/VectorUtility.cs
@@ -7,12 +7,10 @@ namespace Editor.MonoBehaviour
     {
         public static Vector3 Truncate(this Vector3 vec, Vector3 offset = new Vector3())
         {
-
             return new Vector3(
-
-                (int)Math.Truncate(vec.x) + offset.x,
-                (int)Math.Truncate(vec.y) + offset.y,
-                (int)Math.Truncate(vec.z) + offset.z
+                (float) Math.Truncate(vec.x) + offset.x,
+                (float) Math.Truncate(vec.y) + offset.y,
+                (float) Math.Truncate(vec.z) + offset.z
             );
         }
     }
